Fix comments and one resulting code confusion about the type of the "command" argument to VOP_IOCTL.
Spotted by:	FlexeLint.


git-svn-id: svn+ssh://svn.freebsd.org/base/head@105212 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/fs/devfs/devfs_vnops.c
@@ -264,7 +264,7 @@ static int
 devfs_ioctl(ap)
 	struct vop_ioctl_args /* {
 		struct vnode *a_vp;
-		int  a_command;
+		u_long  a_command;
 		caddr_t  a_data;
 		int  a_fflag;
 		struct ucred *a_cred;
